+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkVariable/NetworkVariableCollectionsChangingTests.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: adde18d8af294e4786577a51e48ef109
+timeCreated: 1752086183
